textprocessor/requirements.txt
breadability==0.1.20
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
djangorestframework==3.8.2
docopt==0.6.2
httpie==0.9.9
idna==2.7
lxml==4.2.1
nltk==3.3
numpy==1.14.5
Pygments==2.2.0
pytz==2018.4
requests==2.19.1
six==1.11.0
sumy==0.7.0
urllib3==1.23
uWSGI==2.0.17
